Added a new behavior which allows the user to create new TODO items.This is to demonstrate behaviors and directives can be combined to respond to user interaction.
Along with introduced a new aspect of ng-model directive wherein when supplied the name of a property, for the directive to update, that is not part a model, the property gets dynamically created within the scope of the controller.

Also introduced the ng-click directive which sets up a handler that evaluates the expression when the "click" event is triggered.